add early stop in DSE & add result log recovery

+++ gpu_framework/constants.py
@@ -189,14 +189,20 @@ if not debug and not generate_all_dataset:
         #     file_dir = f"gpu_{mode}/result/{benchmark_name}_{mode}_{lr}_{bs}_{num_epoch}_{train_size}_{use_smooth_kernel}_{num_components}_{l}_{b}_{nn_mode}_{module}_{n}_{save}_{safe_range_list}_{safe_range_bound_list}_{phi_list}_{w_list}.txt"
         #     file_dir_evaluation = f"gpu_{mode}/result/{benchmark_name}_{mode}_{lr}_{bs}_{num_epoch}_{train_size}_{use_smooth_kernel}_{num_components}_{l}_{b}_{nn_mode}_{module}_{n}_{save}_{safe_range_list}_{safe_range_bound_list}_{phi_list}_{w_list}_evaluation.txt"
     
-    log_file = open(file_dir, 'w')
+    if os.path.exists(file_dir):
+        log_file = open(file_dir, 'a')
+    else:
+        log_file = open(file_dir, 'w')
     log_file.write(f"{args}\n")
     log_file.write(f"Target info: {safe_range_list}, {phi_list}, \
         {w_list}, {method_list}, {safe_range_bound_list}\n")
     log_file.write(f"path_num_list: {path_num_list}")
     log_file.close()
 
-    log_file_evaluation =  open(file_dir_evaluation, 'w')
+    if os.path.exists(file_dir):
+        log_file_evaluation = open(file_dir_evaluation, 'a')
+    else:
+        log_file_evaluation =  open(file_dir_evaluation, 'w')
     log_file_evaluation.write(f"{args}\n")
     log_file_evaluation.write(f"Target info: {safe_range_list}, {phi_list}, \
         {w_list}, {method_list}, {safe_range_bound_list}\n")
